fix: change cleaning_cost column type from integer to float in printers table

+++ database/migrations/1736262736589_create_printers_table.ts
@@ -13,7 +13,7 @@ export default class extends BaseSchema {
       table.float("width").notNullable();
       table.float("length").notNullable();
       table.float("height").notNullable();
-      table.integer("cleaning_cost").notNullable();
+      table.float("cleaning_cost").notNullable();
       table.float("impressing_cost").notNullable();
     });
   }
